// Bottom-SPA/src/app/_core/_models/kanban-by-po-detail-by-receiving-type.ts
export interface KanbanByPoDetailByReceivingType {
    plan_Start_STF: string | null;
    line_ID: string;
    mO_No: string;
    mO_Seq: string;
    model_Name: string;
    article: string;
    plan_Qty: number;
    transaction_Date: string | null;
    t2_Supplier: string;
    rack_Location: string;
    material_No: string;
    material_Name: string;
    qrCode_ID: string;
    receiving_Type: string;
    total_Pairs: number | null;
}